refactor: modularize analytics and header controls
