+++ src/App.jsx
@@ -9,8 +9,6 @@ import React, {
 import { useCsvFiles } from './hooks/useCsvFiles';
 import { useSessionManager } from './hooks/useSessionManager';
 import {
-  clusterApneaEvents,
-  detectFalseNegatives,
   FALSE_NEG_CONFIDENCE_MIN,
   FLG_BRIDGE_THRESHOLD,
   APOEA_CLUSTER_MIN_TOTAL_SEC,
@@ -20,7 +18,6 @@ import {
   FLG_EDGE_ENTER_THRESHOLD_DEFAULT,
   FLG_EDGE_EXIT_THRESHOLD_DEFAULT,
 } from './utils/clustering';
-import { finalizeClusters } from './utils/analytics';
 import Overview from './components/Overview';
 const SummaryAnalysis = lazy(() => import('./components/SummaryAnalysis'));
 const ApneaClusterAnalysis = lazy(
@@ -38,9 +35,14 @@ import DocsModal from './components/DocsModal';
 import DataImportModal from './components/DataImportModal';
 import GuideLink from './components/GuideLink';
 import HeaderMenu from './components/HeaderMenu';
+import DateRangeControls from './components/DateRangeControls';
 import { buildSummaryAggregatesCSV, downloadTextFile } from './utils/export';
 import { clearLastSession } from './utils/db';
 import { DataProvider } from './context/DataContext';
+import { useAnalyticsProcessing } from './hooks/useAnalyticsProcessing';
+import { useDateRangeFilter } from './hooks/useDateRangeFilter';
+import { useGuide } from './hooks/useGuide';
+import { useModal } from './hooks/useModal';
 
 function App() {
   const tocSections = useMemo(
@@ -72,16 +74,20 @@ function App() {
     setDetailsData,
     error,
   } = useCsvFiles();
-  const [apneaClusters, setApneaClusters] = useState([]);
-  const [falseNegatives, setFalseNegatives] = useState([]);
-  const [processingDetails, setProcessingDetails] = useState(false);
-  const [dateFilter, setDateFilter] = useState({ start: null, end: null });
-  const [quickRange, setQuickRange] = useState('all');
+  const {
+    dateFilter,
+    setDateFilter,
+    quickRange,
+    handleQuickRangeChange,
+    parseDate,
+    formatDate,
+    selectCustomRange,
+    resetDateFilter,
+  } = useDateRangeFilter(summaryData);
   const [rangeA, setRangeA] = useState({ start: null, end: null });
   const [rangeB, setRangeB] = useState({ start: null, end: null });
-  const [guideOpen, setGuideOpen] = useState(false);
-  const [guideAnchor, setGuideAnchor] = useState('');
-  const [importOpen, setImportOpen] = useState(true);
+  const { isOpen: importOpen, open: openImportModal, close: closeImportModal } =
+    useModal(true);
   const [clusterParams, setClusterParams] = useState({
     gapSec: APNEA_GAP_DEFAULT,
     bridgeThreshold: FLG_BRIDGE_THRESHOLD,
@@ -117,17 +123,12 @@ function App() {
     };
     return presets[fnPreset] || presets.balanced;
   }, [fnPreset]);
-  const latestDate = useMemo(() => {
-    if (!summaryData || !summaryData.length) return new Date();
-    const dateCol = Object.keys(summaryData[0]).find((c) => /date/i.test(c));
-    if (!dateCol) return new Date();
-    return (
-      summaryData.reduce((max, r) => {
-        const d = new Date(r[dateCol]);
-        return !max || d > max ? d : max;
-      }, null) || new Date()
-    );
-  }, [summaryData]);
+  const { apneaClusters, falseNegatives, processing } = useAnalyticsProcessing(
+    detailsData,
+    clusterParams,
+    fnOptions,
+  );
+  const { guideOpen, guideAnchor, openGuideForActive, closeGuide } = useGuide(activeId);
   const { handleLoadSaved, handleExportJson, importSessionFile } =
     useSessionManager({
       summaryData,
@@ -171,112 +172,6 @@ function App() {
   const onClusterParamChange = useCallback((patch) => {
     setClusterParams((prev) => ({ ...prev, ...patch }));
   }, []);
-
-  // Map in-app sections to guide anchors
-  const guideMap = {
-    overview: 'overview-dashboard',
-    'usage-patterns': 'usage-patterns',
-    'ahi-trends': 'ahi-trends',
-    'pressure-settings': 'pressure-correlation-epap',
-    'apnea-characteristics': 'apnea-event-characteristics-details-csv',
-    'clustered-apnea': 'clustered-apnea-events-details-csv',
-    'false-negatives': 'potential-false-negatives-details-csv',
-    'raw-data-explorer': 'raw-data-explorer',
-    'range-compare': 'range-comparisons-a-vs-b',
-  };
-  const openGuideForActive = () => {
-    const anchor = guideMap[activeId] || '';
-    setGuideAnchor(anchor);
-    setGuideOpen(true);
-  };
-
-  // Global event to open guide from inline links without prop drilling
-  useEffect(() => {
-    const handler = (e) => {
-      const anchor = e?.detail?.anchor || '';
-      setGuideAnchor(anchor);
-      setGuideOpen(true);
-    };
-    window.addEventListener('open-guide', handler);
-    return () => window.removeEventListener('open-guide', handler);
-  }, []);
-
-  useEffect(() => {
-    if (detailsData) {
-      // begin processing phase
-      setProcessingDetails(true);
-      // defer clustering/detection to next tick so UI can update (e.g., hide parse progress)
-      let cancelled = false;
-      let worker;
-      try {
-        // eslint-disable-next-line no-undef
-        worker = new Worker(
-          new URL('./workers/analytics.worker.js', import.meta.url),
-          { type: 'module' },
-        );
-        worker.onmessage = (evt) => {
-          if (cancelled) return;
-          const { ok, data, error } = evt.data || {};
-          if (ok) {
-            const clusters = data.clusters || [];
-            setApneaClusters(clusters);
-            setFalseNegatives(data.falseNegatives || []);
-            setProcessingDetails(false);
-          } else {
-            console.warn('Analytics worker error:', error);
-            fallbackCompute();
-          }
-        };
-        worker.postMessage({
-          action: 'analyzeDetails',
-          payload: { detailsData, params: clusterParams, fnOptions },
-        });
-      } catch (err) {
-        console.warn('Worker unavailable, using fallback', err);
-        fallbackCompute();
-      }
-      function fallbackCompute() {
-        const apneaEvents = detailsData
-          .filter((r) =>
-            ['ClearAirway', 'Obstructive', 'Mixed'].includes(r['Event']),
-          )
-          .map((r) => ({
-            date: new Date(r['DateTime']),
-            durationSec: parseFloat(r['Data/Duration']),
-          }));
-        const flgEvents = detailsData
-          .filter((r) => r['Event'] === 'FLG')
-          .map((r) => ({
-            date: new Date(r['DateTime']),
-            level: parseFloat(r['Data/Duration']),
-          }));
-        const rawClusters = clusterApneaEvents(
-          apneaEvents,
-          flgEvents,
-          clusterParams.gapSec,
-          clusterParams.bridgeThreshold,
-          clusterParams.bridgeSec,
-          clusterParams.edgeEnter,
-          clusterParams.edgeExit,
-          10,
-          clusterParams.minDensity,
-        );
-        const validClusters = finalizeClusters(rawClusters, clusterParams);
-        setApneaClusters(validClusters);
-        setFalseNegatives(detectFalseNegatives(detailsData, fnOptions));
-        setProcessingDetails(false);
-      }
-      return () => {
-        cancelled = true;
-        try {
-          worker && worker.terminate && worker.terminate();
-        } catch {
-          // ignore termination errors
-        }
-        setProcessingDetails(false);
-      };
-    }
-  }, [detailsData, clusterParams, fnOptions]);
 
   const filteredSummary = useMemo(() => {
     if (!summaryData) return null;
@@ -377,37 +272,6 @@ function App() {
     // Re-run when data presence changes which may mount/unmount sections
   }, [filteredSummary, filteredDetails, tocSections]);
 
-  const parseDate = (val) => {
-    if (!val) return null;
-    const d = new Date(val);
-    return Number.isNaN(d.getTime()) ? null : d;
-  };
-
-  const handleQuickRangeChange = useCallback(
-    (val) => {
-      setQuickRange(val);
-      if (val === 'all') {
-        setDateFilter({ start: null, end: null });
-      } else if (val !== 'custom') {
-        const days = parseInt(val, 10);
-        if (!Number.isNaN(days)) {
-          const end = latestDate;
-          const start = new Date(end);
-          start.setDate(start.getDate() - (days - 1));
-          setDateFilter({ start, end });
-        }
-      }
-    },
-    [latestDate],
-  );
-
-  const formatDate = (d) =>
-    d instanceof Date && !Number.isNaN(d.getTime())
-      ? new Date(d.getTime() - d.getTimezoneOffset() * 60000)
-          .toISOString()
-          .slice(0, 10)
-      : '';
-
   return (
     <DataProvider
       summaryData={summaryData}
@@ -419,13 +283,13 @@ function App() {
     >
       <DataImportModal
         isOpen={importOpen}
-        onClose={() => setImportOpen(false)}
+        onClose={closeImportModal}
         onSummaryFile={onSummaryFile}
         onDetailsFile={onDetailsFile}
         onLoadSaved={handleLoadSaved}
         onSessionFile={importSessionFile}
         loadingSummary={loadingSummary}
-        loadingDetails={loadingDetails || processingDetails}
+        loadingDetails={loadingDetails || processing}
         summaryProgress={summaryProgress}
         summaryProgressMax={summaryProgressMax}
         detailsProgress={detailsProgress}
@@ -450,64 +314,20 @@ function App() {
             <h1>OSCAR Sleep Data Analysis</h1>
             <span className="badge">beta</span>
           </div>
-          <div className="date-filter">
-            <select
-              value={quickRange}
-              onChange={(e) => handleQuickRangeChange(e.target.value)}
-              aria-label="Quick range"
-            >
-              <option value="all">All</option>
-              <option value="7">Last 7 days</option>
-              <option value="14">Last 14 days</option>
-              <option value="30">Last 30 days</option>
-              <option value="90">Last 90 days</option>
-              <option value="180">Last 180 days</option>
-              <option value="365">Last year</option>
-              <option value="1825">Last 5 years</option>
-              <option value="custom">Custom</option>
-            </select>
-            <input
-              type="date"
-              value={formatDate(dateFilter.start)}
-              onChange={(e) => {
-                setQuickRange('custom');
-                setDateFilter((prev) => ({
-                  ...prev,
-                  start: parseDate(e.target.value),
-                }));
-              }}
-              aria-label="Start date"
-            />
-            <span>-</span>
-            <input
-              type="date"
-              value={formatDate(dateFilter.end)}
-              onChange={(e) => {
-                setQuickRange('custom');
-                setDateFilter((prev) => ({
-                  ...prev,
-                  end: parseDate(e.target.value),
-                }));
-              }}
-              aria-label="End date"
-            />
-            {(dateFilter.start || dateFilter.end) && (
-              <button
-                className="btn-ghost"
-                onClick={() => {
-                  setDateFilter({ start: null, end: null });
-                  setQuickRange('all');
-                }}
-                aria-label="Reset date filter"
-              >
-                ×
-              </button>
-            )}
-          </div>
+          <DateRangeControls
+            quickRange={quickRange}
+            onQuickRangeChange={handleQuickRangeChange}
+            dateFilter={dateFilter}
+            onDateFilterChange={setDateFilter}
+            onCustomRange={selectCustomRange}
+            onReset={resetDateFilter}
+            parseDate={parseDate}
+            formatDate={formatDate}
+          />
           <div className="actions">
             <ThemeToggle />
             <HeaderMenu
-              onOpenImport={() => setImportOpen(true)}
+              onOpenImport={openImportModal}
               onExportJson={handleExportJson}
               onExportCsv={exportAggregatesCsv}
               onClearSession={handleClearSession}
@@ -518,7 +338,7 @@ function App() {
             />
           </div>
         </div>
-        {(loadingSummary || loadingDetails || processingDetails) && (
+        {(loadingSummary || loadingDetails || processing) && (
           <div className="import-progress" aria-live="polite">
             <span>
               {loadingSummary &&
@@ -538,7 +358,7 @@ function App() {
                       )}%)`
                     : ''
                 }`}
-              {processingDetails &&
+              {processing &&
                 !loadingSummary &&
                 !loadingDetails &&
                 'Processing events...'}
@@ -722,7 +542,7 @@ function App() {
         )}
         <DocsModal
           isOpen={guideOpen}
-          onClose={() => setGuideOpen(false)}
+          onClose={closeGuide}
           initialAnchor={guideAnchor}
         />
       </div>
