+++ src/hooks/useModal.js
@@ -1,0 +1,10 @@
+import { useCallback, useState } from 'react';
+
+export function useModal(initial = false) {
+  const [isOpen, setIsOpen] = useState(initial);
+
+  const open = useCallback(() => setIsOpen(true), []);
+  const close = useCallback(() => setIsOpen(false), []);
+
+  return { isOpen, open, close };
+}
